+++ src/GetSet/Delete.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Parable\GetSet;
+
+class Delete extends \Parable\GetSet\BaseInput
+{
+    /** @var string */
+    protected $resource = '_DELETE';
+}