+++ src/GetSet/Patch.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Parable\GetSet;
+
+class Patch extends \Parable\GetSet\BaseInput
+{
+    /** @var string */
+    protected $resource = '_PATCH';
+}