+++ src/GetSet/Put.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Parable\GetSet;
+
+class Put extends \Parable\GetSet\BaseInput
+{
+    /** @var string */
+    protected $resource = '_PUT';
+}